fix(FEC-9423): Small view on desktop the pause icon remains on screen
it'll work on ad which playing on same video tag cause we'll get the correct state of paused.
On different video tags we are getting the state of paused from the main video tag but ad is playing on second one and pause icon still showing so added state of adBreak for this situtation.