fix(ADA-1275): When user focuses the player he will be able to play/pause with "space" key (#907)
**Issue:**
After the pre-playback-overlay is rendered, with the centered big play
button shown in the player container, the user can use it to start the
playback and make the overlay component disappear. Afterwards he can
play/pause the media by pressing "Space" key due to the fact that
initially the play/pause button is focused.

The issue appears when the user switches focus from the player and
returns to it, because the play button will not be focused anymore.

**Fix:**

Added event listener on the player container for the space key.