fix(SUP-46391): Player Live Full Screen Bug when browser window is not maximized (#976)
Issue: 
When browser width is small, side pannel is on, and enter full screen,
there is a gap between side pannel and player.

Fix:
Fix timing issue : the issue related to the timing of the DOM updates.
When resizing the window or triggering an actions that affect the
layout, the client rect of the player container getBoundingClientRect()
is not updated immediately after the resize. This causes discrepancies
between the actual layout and the values being used in the state.

solved [SUP-46391](https://kaltura.atlassian.net/browse/SUP-46391)

[SUP-46391]:
https://kaltura.atlassian.net/browse/SUP-46391?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ