fix (SUP-47338): Watermark On Player Not Responsive (#997)
**Issue:**
Currently when client applies a watermark image, it is just applied on
the player "as is" without calculate the proportion considering the
player height and width.

**Fix:**
Calculate the aspect ratio of the image, and apply new height and width
based on the player size.
Listen to resize event and update the image size whenever resize is
being dispatched.

solved https://kaltura.atlassian.net/browse/SUP-47338