fix(ADA-2122): Include modal title in close button aria-label (#988)
This solves this https://kaltura.atlassian.net/browse/ADA-2122. It
includes the modal title to the Advanced captions settings modal's close
button.
This should be merged after this
https://github.com/kaltura/playkit-js-ui/pull/984

### Description of the Changes

Please add a detailed description of the change, whether it's an
enhancement or a bugfix.
If the PR is related to an open issue please link to it.

**Issue:**

**Fix:**

#### Resolves FEC-[Please add the ticket reference here]