fix(ADA-2432): Focus on settings button on close (#989)
This solves this https://kaltura.atlassian.net/browse/ADA-2432.
The issue appeared because the captions menu and settings menu are
closing when the Advanced captions settings is selected. Now on modal
close, the focus searches for Settings button class and focuses on it.

### Description of the Changes

Please add a detailed description of the change, whether it's an
enhancement or a bugfix.
If the PR is related to an open issue please link to it.

**Issue:**

**Fix:**

#### Resolves FEC-[Please add the ticket reference here]