fix (SUP-47965): Hotkey F is not working when the focus is on another button in the video. (#1016)
**Issue:**
1. When focus on a button in the player using tab, and press F, the
player doesn't enter to fullscreen mode.
2. Hotkey F is working only for one direction when the focus is on the
player without using tab:
when press on F when the player is not on fullscreen mode, makes the
player enter to fullscreen mode but not on the oposite: when player is
on fullscreen mode, and press F, nothing happens.

**Fix:**
support Hotkey F when player fullscreen mode is on for both in shells
and using tab.

solved [SUP-47965](https://kaltura.atlassian.net/browse/SUP-47965)

[SUP-47965]:
https://kaltura.atlassian.net/browse/SUP-47965?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ