fix(ADA-2821) - Make apply button keyboard accessible (#1042)
This solves https://kaltura.atlassian.net/browse/ADA-2821 making Apply
button keyboard accessible. After apply, the focus should go to the
button that opened the modal
### Description of the Changes

Please add a detailed description of the change, whether it's an
enhancement or a bugfix.
If the PR is related to an open issue please link to it.

**Issue:**

**Fix:**

#### Resolves FEC-[Please add the ticket reference here]