fix(ADA-2781): Pre playback play focus management (#1029)
This solves https://kaltura.atlassian.net/browse/ADA-2781 by making
focus go on the Play/Pause button after pre-playback play button click

### Description of the Changes

Please add a detailed description of the change, whether it's an
enhancement or a bugfix.
If the PR is related to an open issue please link to it.

**Issue:**

**Fix:**

#### Resolves FEC-[Please add the ticket reference here]