fix(SUP-49217): [Minn State] Cannot select Download Pre-Test in Video… (#1064)
Quiz based on YouTube entry in v7 player

Description of the Changes
Issue:
In case of youtube + quiz entry, the Download Pre-Test button is not clickable

Root cause:
The prePlaybackOverlay is covering the quiz overlay (it has z-index: 2)

Fix:
Add the z-index: 2 only if it's not a quiz entry.
In quiz entries, it is not possible (also on non youtube entries) to play the video by clicking on the overlay, only on the play button.

Resolves SUP-49217