fix(ADA-2934): Remove roles from native select (#1073)
This solves https://kaltura.atlassian.net/browse/ADA-2934. The native
select does not need ARIA attributes to work correctly. The SR was
confused by the roles used before.

### Description of the Changes

Please add a detailed description of the change, whether it's an
enhancement or a bugfix.
If the PR is related to an open issue please link to it.

**Issue:**

**Fix:**

#### Resolves FEC-[Please add the ticket reference here]