fix(ADA-2938): Change dropdown role (#1072)
This solves https://kaltura.atlassian.net/browse/ADA-2938. Taking
https://www.w3.org/WAI/ARIA/apg/patterns/combobox/examples/combobox-select-only/
in consideration, I changed the actual dropdown role to combobox and the
role of the menuitem to option.

### Description of the Changes

Please add a detailed description of the change, whether it's an
enhancement or a bugfix.
If the PR is related to an open issue please link to it.

**Issue:**

**Fix:**

#### Resolves FEC-[Please add the ticket reference here]