fix(ADA-1623): OnClose focus on plugin button (#1078)
This pr solves this https://kaltura.atlassian.net/browse/ADA-1623. It
returns the focus on the plugin button after close. _setPluginButtonRef
was never running, because of wrapping the button with <Toolbar>. The
problem was that Tooltip was overwriting the child’s ref. I updated
Tooltip to merge refs, so it keeps its own reference and still forwards
the child’s.

### Description of the Changes

Please add a detailed description of the change, whether it's an
enhancement or a bugfix.
If the PR is related to an open issue please link to it.

**Issue:**

**Fix:**

#### Resolves FEC-[Please add the ticket reference here]