fix(SUP-45260): University of California - Berkeley | v7 player issue when using picture-in-picture (#1079)
Description of the Changes
Issue:
Picture in picture feature is not working on dual screen players especially when it's on top bar.

Fix:
Adding an API that stores the the dual screen players then use it in picture in picture component in order to enter to pip with the right player.

Resolves UP-45260