fix(ADA-2709): Remove Advanced captions settings from caption list (#1086)
This solves https://kaltura.atlassian.net/browse/ADA-2709 by removing
Advanced captions settings button from captions list and separating the
additional options by normal options with a line as per desing
https://www.figma.com/design/8yTuCnuVHLYTFp89w7z05J/%F0%9F%93%BA-Player-v7?node-id=20974-30781&m=dev

### Description of the Changes

Please add a detailed description of the change, whether it's an
enhancement or a bugfix.
If the PR is related to an open issue please link to it.

**Issue:**

**Fix:**

#### Resolves FEC-[Please add the ticket reference here]