fix(ADA-2976): Prevent select to open on mount (#1083)
This solves this https://kaltura.atlassian.net/browse/ADA-2976.
I could not find a solution to fix this using only preact life cycle
methods. requestAnimationFrame runs code right before the browser paints
the next frame, and it’s part of the official rendering lifecycle.
Deferring focus this way makes sure the DOM is ready. For <select>
elements, it prevents browsers from auto-opening on mount.


### Description of the Changes

Please add a detailed description of the change, whether it's an
enhancement or a bugfix.
If the PR is related to an open issue please link to it.

**Issue:**

**Fix:**

#### Resolves FEC-[Please add the ticket reference here]