fix(ADA-2781): Play buttons focus management (#1059)
This solves the entering fullscreen mode on play/pause button
https://kaltura.atlassian.net/browse/ADA-2781.
This includes the code used
https://github.com/kaltura/playkit-js-ui/pull/1029/files that created
the black line regression. I am not seeing the black line locally.

### Description of the Changes

Please add a detailed description of the change, whether it's an
enhancement or a bugfix.
If the PR is related to an open issue please link to it.

**Issue:**

**Fix:**

#### Resolves FEC-[Please add the ticket reference here]