fix(ADA-2858): Advanced captions settings feature (#1080)
This solves https://kaltura.atlassian.net/browse/ADA-2858 updating the
Advanced captions settings format.
It is related to https://github.com/kaltura/playkit-js/pull/831.
Translations https://kaltura.atlassian.net/browse/FEC-14644
### Description of the Changes

Please add a detailed description of the change, whether it's an
enhancement or a bugfix.
If the PR is related to an open issue please link to it.

**Issue:**

**Fix:**

#### Resolves FEC-[Please add the ticket reference here]