fix(FEC-14228): Switching between regular audio language to Audio description doesn't works properly (#1107)
### Description of the Changes

After a backend fix, the manifest structure now supports switching
between audio track and audio description track for the same language.

What was missing for the feature to work is a fix on the ui side where
when switching from the audio description track to the audio track, we
weren't using the language field from the audio description track, so
the switch didn't work on the ui side.

This fix resolves this issue.

Resolves FEC-14228