fix(FEC-14684): Advanced captions settings feature (#1108)
This solves https://kaltura.atlassian.net/browse/FEC-14684 and should be
treated as a feature. It changes the Advanced captions settings
structure.
Figma file:
https://www.figma.com/design/8yTuCnuVHLYTFp89w7z05J/%F0%9F%93%BA-Player-v7?node-id=22518-53428&t=CW2rUDmwnZW3Ky0x-1

Relates to https://github.com/kaltura/playkit-js/pull/840 and
https://github.com/kaltura/playkit-js-kava/pull/214 and should be merged
together

### Description of the Changes

Please add a detailed description of the change, whether it's an
enhancement or a bugfix.
If the PR is related to an open issue please link to it.

**Issue:**

**Fix:**

#### Resolves FEC-[Please add the ticket reference here]