Revert "fix(ADA-2709): Remove Advanced captions settings from caption… (#1116)
Because here https://kaltura.atlassian.net/browse/ADA-2709 the decision
that we should not the the ticket was made, I am reverting the changes
made https://github.com/kaltura/playkit-js-ui/pull/1086, except the
visual border that separates Advanced captions settings from the rest of
the list.

### Description of the Changes

Please add a detailed description of the change, whether it's an
enhancement or a bugfix.
If the PR is related to an open issue please link to it.

**Issue:**

**Fix:**

#### Resolves FEC-[Please add the ticket reference here]