fix(FEC-14573): The gap between player controls should be 8px (#1123)
continue of - #1117

Issue:
In player resizing, time is removed and added according to the space in the bottom bar(for example when open a panel side like transcript) then padding style is not added correctly.

Root cause:
In the above case the time-display component is re-initialized, so componentDidUpdate hasn’t been called yet.

Fix:
Add functionality both in componentDidMount and componentDidUpdate (need both cause in the first time the state is updated after the the initial render)

Resolves FEC-14573