fix(FEC-14813): Change minimalist style (#1134)
This solves https://kaltura.atlassian.net/browse/FEC-14813 by making the
minimalist styling equal to the default styling of the captions. Related
to this revert https://github.com/kaltura/playkit-js/pull/850
### Description of the Changes

Please add a detailed description of the change, whether it's an
enhancement or a bugfix.
If the PR is related to an open issue please link to it.

**Issue:**

**Fix:**

#### Resolves FEC-[Please add the ticket reference here]